Move to type qualifier to type utils
